Rename: Rename Profiles Page
Rename profile/* to profiles/* as it shows multiple files.
Update <a> in index menu.
