+++ src/index.html
@@ -9,7 +9,7 @@
 	<body>
 		<h1>VMS pages</h1>
 		<ul>
-			<li><a href="/pages/profile/">Profile Page</a></li>
+			<li><a href="/pages/profiles/">Profiles Page</a></li>
 			<li><a href="/pages/team/">Team Page</a></li>
 		</ul>
 	</body>
