+++ src/pages/profiles/index.html
@@ -3,10 +3,10 @@
 	<head>
 		<meta charset="UTF-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Profile</title>
+		<title>Profiles</title>
 	</head>
 	<body>
 		<div id="root"></div>
-		<script type="module" src="./profile.tsx"></script>
+		<script type="module" src="./profiles.tsx"></script>
 	</body>
 </html>
